+++ src/Installer/InstallerFactory.php
@@ -74,6 +74,11 @@ class InstallerFactory
     public $noInit;
 
     /**
+     * @var string
+     */
+    public $nodeVer;
+
+    /**
      * Given a big bag of install options, add installers to the collection.
      *
      * @param InstallerCollection $installers Installers will be added to this
@@ -82,7 +87,7 @@ class InstallerFactory
     {
         $installers->add(new MoodleInstaller($this->execute, $this->database, $this->moodle, new MoodleConfig(), $this->repo, $this->branch, $this->dataDir));
         $installers->add(new PluginInstaller($this->moodle, $this->plugin, $this->pluginsDir, $this->dumper));
-        $installers->add(new VendorInstaller($this->moodle, $this->plugin, $this->execute));
+        $installers->add(new VendorInstaller($this->moodle, $this->plugin, $this->execute, $this->nodeVer));
 
         if ($this->noInit) {
             return;